sql bug

+++ src/main/java/life/my/community/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer viewCount;
+    private Long viewCount;
     private Integer creator;
     private Integer commentCount;
     private Integer likeCount;
